requirements.txt
flask
python-dotenv
requests
pytest
pytest-cov
flake8
black
flask-talisman
